requirements.txt
openai
langchain
langchain_openai
streamlit
python-dotenv
PyPDF2
openpyxl